+++ src/ContinuousStepper_Tone.h
@@ -1,37 +1,16 @@
 #pragma once
 
-#include "ContinuousStepperBase.h"
+#include "ContinuousStepperWithOscillator.hpp"
 
 namespace ArduinoContinuousStepper {
-class ContinuousStepper_Tone : public ContinuousStepperBase {
-public:
-  void loop() {
-    if (!_period)
-      return;
-
-    time_t now = micros();
-    time_t elapsed = now - _periodStart;
-
-    if (elapsed >= _period) {
-      updateSpeedIfNeeded();
-      _periodStart = now;
-    }
+class ContinuousStepper_Tone : public ContinuousStepperWithOscillator {
+  void startOscillator(unsigned int frequency) override {
+    tone(stepPin(), frequency);
   }
 
-private:
-  void setPeriod(unsigned long period) override {
-    if (period) {
-      writeDir();
-      tone(stepPin(), 1e6 / period);
-    } else {
-      noTone(stepPin());
-    }
-    _period = period;
+  void stopOscillator() override {
+    noTone(stepPin());
   }
-
-  void writeStep(bool) override {}
-
-  time_t _period, _periodStart = 0;
 };
 } // namespace ArduinoContinuousStepper