+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Fix occasional missed ticks on timer-based implementations
 * Add `ContinuousStepper_Tone` which uses `tone()` instead of `digitalWrite()` for the step pin
+* Add `ContinuousStepper_AWF` which uses `analogWriteFrequency()` on Teensy 3 and 4
 
 2.1.0
 -----